+++ packages/antdv/components/collapse-transtion/index.ts
@@ -1,0 +1,7 @@
+import { withInstall } from '../utils/install'
+import CollapseTranstion from './src/collapse-transtion.vue'
+import './style'
+
+export * from './src/types'
+export const KCollapseTranstion = withInstall(CollapseTranstion)
+export default KCollapseTranstion