+++ packages/antdv/components/index.ts
@@ -4,7 +4,8 @@
  * @Autor: caohao
  * @Date: 2024-09-17 00:26:48
  * @LastEditors: error: error: git config user.name & please set dead value or install git && error: git config user.email & please set dead value or install git & please set dead value or install git
- * @LastEditTime: 2024-10-15 15:49:07
+ * @LastEditTime: 2024-10-21 18:03:06
  */
 export * from './form'
 export * from './virtual-list'
+export * from './collapse-transtion'